requirements.txt
langchain
langchain-openai
langchain-community
python-dotenv
fastapi
uvicorn
requests
pydantic